Reviewing and separating the function
