+++ Makefile
@@ -5,7 +5,14 @@ LFLAGS = -lreadline
 RM = rm -f
 INCLUDES = -I./includes/ -I./libft/includes/
 
-SRCS = main.c ./parsing/parsing.c ./parsing/tokenizing.c
+SRCS = main.c \
+	./tokenizing/tokenizing.c ./tokenizing/clean_tokens.c \
+	./tokenizing/print_tokens.c ./tokenizing/tokenizing_utils.c \
+	./parsing/parsing.c ./parsing/print_ast.c \
+	./parsing/free_ast.c ./parsing/build_ast.c \
+	./parsing/create_command.c ./parsing/redirections.c \
+
+
 OBJS = $(SRCS:.c=.o)
 
 LIBFT = ./libft/libft.a
