+++ tokenizing/tokenizing.c
@@ -1,0 +1,40 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   tokenizing.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: javokhir <javokhir@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/08/27 22:19:04 by javokhir          #+#    #+#             */
+/*   Updated: 2025/08/28 08:54:36 by javokhir         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/minishell.h"
+
+t_token	*ft_tokenize(char *s)
+{
+	t_token	*list;
+	int		i;
+
+	list = NULL;
+	i = 0;
+	while (s[i])
+	{
+		while (s[i] == ' ' || (s[i] >= '\t' && s[i] <= '\r'))
+			i++;
+		if (!s[i])
+			break ;
+		if (ft_strchr("|><", s[i]) && s[i + 1] && s[i + 1] == s[i])
+			i += specify_tokens(s[i], &list, 0);
+		else if (ft_strchr("|><", s[i]) && !(s[i + 1] && s[i + 1] == s[i]))
+			i += specify_tokens(s[i], &list, 1);
+		else if (ft_strchr("'\"", s[i]))
+			i += handle_quotes(s + i, s[i], &list);
+		else
+			i += handle_words(s + i, &list);
+	}
+	add_tokens(&list, T_EOF, NULL);
+//	print_tokens(list);//should be removed later
+	return (list);
+}
